+++ demos/examples/buffers/image.glsl
@@ -1,0 +1,4 @@
+void mainImage(out vec4 fragColor, in vec2 fragCoord)
+{
+    fragColor = texelFetch(state, ivec2(fragCoord), 0);
+}