+++ demos/quickstart/03-one-buffer/image.glsl
@@ -1,6 +1,0 @@
-// Image: Display what BufferA rendered
-
-void mainImage(out vec4 fragColor, in vec2 fragCoord)
-{
-    fragColor = texelFetch(iChannel0, ivec2(fragCoord), 0);
-}